[Fix] AWS에서 기본사진 날라간 것 수정

+++ backend/src/main/java/hello/wouldyoulist/controller/ReviewController.java
@@ -133,44 +133,35 @@ public class ReviewController {
         reviewService.updateReview(reviewId, request.getParameter("title"), request.getParameter("review"), request.getParameter("doneDate"),
                 request.getParameter("place"), request.getParameter("expression")); //커맨드(수정)와
         Review findReview = reviewService.findOne(reviewId).get(); //쿼리(조회)를 분리
+
         Long photoId = findReview.getPhotoId();
-
-        if(file == null || file.isEmpty()){
+        if(file == null || file.isEmpty()) {
             // Case A: file false && del true
-            if(isDeleted){
-                if (photoId != 1) {
-                    // 기존 이미지 삭제
-                    s3Uploader.deleteFile(fileService.findOne(photoId).get().getFullPath());
-                    fileService.deleteFile(photoId);
-
-                    // 기본 이미지로 변경
-                    reviewService.updateReviewPhotoId(reviewId, 1L);
-                }
+            if(isDeleted && photoId != 1) {
+                // 기존 이미지 삭제
+                s3Uploader.deleteFile(fileService.findOne(photoId).get().getFullPath());
+                fileService.deleteFile(photoId);
+                // 기본 이미지로 변경
+                reviewService.updateReviewPhotoId(reviewId, 1L);
             }
             // Case B: file false && del false
-            // 유지
+            // 기존 이미지 유지
         }
-        else{
+        else
+//            if(isDeleted)
             // Case C: file true && del true
-            // if(isDeleted){
-                if (photoId != 1) {
-                    // 기존 이미지 삭제
-                    s3Uploader.deleteFile(fileService.findOne(photoId).get().getFullPath());
-                    fileService.deleteFile(photoId);
-                }
-                // 새로운 파일 업로드
-                String originalFilename = file.getOriginalFilename();
-                String storedFileName=s3Uploader.upload(file,"images");
-                Long fileId = fileService.save(new UploadFile(originalFilename, storedFileName));
-                reviewService.updateReviewPhotoId(reviewId, fileId); // table에 photoId 업데이트
-            // }
+            if (photoId != 1) {
+                // 기존 이미지 삭제
+                s3Uploader.deleteFile(fileService.findOne(photoId).get().getFullPath());
+                fileService.deleteFile(photoId);
+            }
+            // 새로운 파일 업로드
+            String originalFilename = file.getOriginalFilename();
+            String storedFileName=s3Uploader.upload(file,"images");
+            Long fileId = fileService.save(new UploadFile(originalFilename, storedFileName));
+            reviewService.updateReviewPhotoId(reviewId, fileId); // table에 photoId 업데이트
+//            else
             // Case D: file true && del false 원래는 추가가 맞음. 지금은 사진 하나라 변경점 X
-//            else{
-//
-//
-//            }
-        }
-
 
         return new UpdateReviewResponse(findReview.getId(), findReview.getTitle());
     }
